Add files via upload
Numerically solves a coupled pair of 1st order ODE using Runge-Kutta method. Additionally, the approcximate solution curve is plotted in conjunction with a meshgrid of points and the assocaited vector field (fixed magnitude). Later updates will include a color map for the vector field (to represent magnitude) and 3D plotting options. Furthermore, the RK solver function is not limited low dimensions, and I will at some point add an option to plot projections of solutions in a larger phas-space to any 2D and 3D subspace.